+++ App.jsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from "@react-navigation/stack";
+import { createStackNavigator, StackView } from '@react-navigation/stack';
 import HomeTabs from './src/navigators/index';
-import Login from './src/login/index';
+import Login from './src/login';
 import Cadastro from './src/cadastro/index';
 import Caduser from './src/cadastro/cadastroUser';
 import CadEmp from './src/cadastro/cadastroEmpresa';
 import BemVindo from './src/login/boasVindas';
 const Stack = createStackNavigator();
-const App: React.FC = () => {
+export default function App() {
   return (
     <NavigationContainer>
       <Stack.Navigator initialRouteName='Login'>
@@ -85,5 +85,4 @@ const App: React.FC = () => {
       </Stack.Navigator>
     </NavigationContainer>
   );
-};
-export default App;
+}
