+++ package.json
@@ -11,34 +11,25 @@
   "dependencies": {
     "@expo/webpack-config": "^18.0.1",
     "@react-native-community/checkbox": "^0.5.16",
-    "@react-native-masked-view/masked-view": "0.2.8",
-    "@react-navigation/bottom-tabs": "^6.5.11",
-    "@react-navigation/native": "^6.1.9",
+    "@react-native-masked-view/masked-view": "^0.2.8",
+    "@react-navigation/bottom-tabs": "^6.5.8",
+    "@react-navigation/native": "^6.1.7",
     "@react-navigation/native-stack": "^6.9.13",
-    "@react-navigation/stack": "^6.3.20",
-    "@types/react": "~18.0.27",
+    "@react-navigation/stack": "^6.3.17",
     "expo": "~48.0.18",
     "expo-linear-gradient": "~12.1.2",
-    "expo-location": "~15.1.1",
-    "expo-permissions": "~14.1.1",
     "expo-status-bar": "~1.4.4",
     "latest": "^0.2.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-native": "0.71.14",
+    "react-native": "0.71.8",
     "react-native-gesture-handler": "~2.9.0",
-    "react-native-maps": "1.3.2",
-    "react-native-navigation": "^7.37.2",
     "react-native-reanimated": "~2.14.4",
     "react-native-snap-carousel": "^3.9.1",
-    "react-native-web": "~0.18.10",
-    "typescript": "^4.9.4"
+    "react-native-web": "~0.18.10"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0",
-    "@types/react": "^18.2.42",
-    "@types/react-native": "^0.72.8",
-    "typescript": "^5.3.2"
+    "@babel/core": "^7.20.0"
   },
   "private": true
 }
