+++ src/mapa/index.jsx
@@ -1,14 +1,11 @@
 import { useNavigation } from '@react-navigation/native';
-import { useRef } from 'react';
 import { StyleSheet, Text, View } from 'react-native';
-import MapView from 'react-native-maps';
 
 export default function Maps() {
   const navigation=useNavigation();
-  const mapEl = useRef(null)
   return (
     <View style={styles.container}>
-     
+      <Text>TESTE Mapa</Text>
     </View>
   );
 }
