--- type.d.ts
@@ -1,4 +1,0 @@
-declare module '*.png';
-declare module '*.jpg';
-declare module '*.jpeg';
-
